Refactor: passage à un paquet installable src/snakeai/
Réorganise le code à plat en paquet Python src/snakeai/ avec séparation
stricte par responsabilité (core, perception, learning, training, ui).

- core/environment.py : board et règles du jeu
- perception/interpreter.py : état (vision) et rewards
- learning/agent.py : Q-learning
- training/trainer.py : boucle d'entraînement extraite du CLI
- ui/display.py + ui/dashboard/ (simulation, renderer, theme, app)
- cli.py : point d'entrée, câblage des composants

Supprime PLAN.pdf et requirements.txt (remplacé par pyproject.toml/uv).
