+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "learn2slither"
 version = "0.1.0"
-description = "Add your description here"
+description = "Snake 10x10 pilote par Q-learning (projet 42)."
 readme = "README.md"
 requires-python = ">=3.14"
 dependencies = [
@@ -9,7 +9,17 @@ dependencies = [
     "pygame-ce>=2.5.7",
 ]
 
+[project.scripts]
+snake = "snakeai.cli:main"
+
 [dependency-groups]
 dev = [
     "flake8>=7.3.0",
 ]
+
+[build-system]
+requires = ["setuptools>=68"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+where = ["src"]
