--- requirements.txt
@@ -1,3 +1,0 @@
-pygame
-numpy
-flake8
