+++ src/snakeai/__main__.py
@@ -1,0 +1,6 @@
+"""Permet `python -m snakeai` : delegue au point d'entree CLI."""
+
+from snakeai.cli import main
+
+if __name__ == "__main__":
+    main()
