+++ src/snakeai/constants.py
