+++ src/snakeai/core/__init__.py
@@ -1,0 +1,5 @@
+"""core - le board et les regles du jeu (aucune logique d'apprentissage)."""
+
+from snakeai.core.environment import Environment
+
+__all__ = ["Environment"]
