+++ src/snakeai/learning/__init__.py
@@ -1,0 +1,5 @@
+"""learning - Q-learning : Q-table, politique epsilon-greedy, update."""
+
+from snakeai.learning.agent import Agent
+
+__all__ = ["Agent"]
