+++ src/snakeai/perception/__init__.py
@@ -1,0 +1,5 @@
+"""perception - vision du serpent (etat) et calcul des rewards."""
+
+from snakeai.perception.interpreter import Interpreter
+
+__all__ = ["Interpreter"]
