+++ src/snakeai/training/__init__.py
@@ -1,0 +1,5 @@
+"""training - boucle d'entrainement assemblant le flux du jeu."""
+
+from snakeai.training.trainer import run_session, train
+
+__all__ = ["run_session", "train"]
