+++ src/snakeai/ui/__init__.py
@@ -1,0 +1,1 @@
+"""ui - affichages pygame optionnels (jamais requis par la logique de jeu)."""
