+++ src/snakeai/ui/dashboard/theme.py
@@ -1,0 +1,23 @@
+"""theme.py - couleurs et metriques de mise en page du dashboard.
+
+Regroupe les constantes purement visuelles pour que la simulation et
+l'orchestration n'aient jamais a connaitre le detail du rendu.
+"""
+
+# Couleurs (R, V, B) - palette minimaliste.
+COLOR_BG = (12, 12, 16)
+COLOR_PANEL = (22, 22, 30)
+COLOR_BOARD_BG = (28, 28, 36)
+COLOR_HEAD = (90, 230, 130)
+COLOR_BODY = (45, 150, 90)
+COLOR_GREEN_APPLE = (70, 200, 70)
+COLOR_RED_APPLE = (215, 70, 70)
+COLOR_TEXT = (220, 220, 230)
+COLOR_DIM = (140, 140, 155)
+COLOR_LEADER = (240, 210, 80)
+COLOR_SELECT = (90, 200, 240)
+
+# Metriques (pixels).
+SIDEBAR = 240
+HEADER = 44
+GAP = 6
